+++ src/ui/scroll-shadower/scroll-shadower.interface.tsx
@@ -1,0 +1,16 @@
+export interface IScrollShadowerProps {
+  height: string
+  width: string
+}
+
+export interface IHorizontalShadowProps {
+  showOnLeft: boolean
+  showOnRight: boolean
+  rightOffset?: number
+}
+
+export interface IVerticalShadowProps {
+  showOnTop: boolean
+  showOnBottom: boolean
+  bottomOffset?: number
+}
